requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
email-validator==2.1.0
reportlab==4.0.7